+++ src/org/espenhahn/allocate/likepaxos/Proposal.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public interface Proposal<E> extends Serializable {
 
-	int getProposalNumber();
+	double getProposalNumber();
 	
 	E getValue();
